faker-producer/requirements.txt
kafka-python
dataprep
faker